Add support for LOCAL_REEXPORTED_SHARED_LIBRARIES
Libraries listed in this variable will work similarly to
LOCAL_SHARED_LIBRARIES, except that headers from those libraries
are exported as part of the current libraries export_includes.

This means that when library A's public headers include headers from
library B, library C which consumes A no longer has to explicitly list
A and B in its LOCAL_SHARED_LIBRARIES for the purpose of picking up B's
headers.

Bug: 27804373
Test: Introduced dependency on libbase to libbinder at the header level.
      Libraries that consume libbinder do not need to explicitly depend
      on libbase.

Change-Id: If69079f058a4a182c5ef5a3c5ba15035570da22d