Update patch set 1
Patch Set 1:

This has been on our todo list for a while.  I was thinking of leaving libbase in LOCAL_SHARED_LIBRARIES, and then using LOCAL_EXPORT_C_INCLUDE_DIRS_FROM := libbase to reexport the headers.  Disadvantage is that it means libbase is listed twice in the Android.mk, advantage is that it works for static libraries too.

Patch-set: 1