fs_config: support parsing android_filesystem_config.h
Rather than hardcode the OEM ranges, parse and extract
AID values from android_filesystem_config.h.

An AID is defined to the tool as a #define AID_<name>

An OEM Range is defined to the the tool like so:
  * AID_OEM_RESERVED_START
  * AID_OEM_RESERVED_END
  * AID_OEM_RESERVED_1_START
  * AID_OEM_RESERVED_1_END

While parsing support sanity checks such as:
1. AIDs defined in the header cannot be within OEM range
2. OEM Ranges must be valid:
   * Cannot overlap one another.
   * Range START must be less than range END
3. Like the C preproccessor, multiple matching AID_<name> throws
   en error.

The parser introduced here, prepares the tool to output android_ids
consumable for bionic.

Change-Id: Iaabdcc46ca8e3b73fa622823479d43cf0f245557
Signed-off-by: William Roberts <william.c.roberts@intel.com>