Update patch set 4
Patch Set 4:

> > Do _you_ have access to gpylint?
 > >
 > > It found a large number of issues even before we started. Most
 > all
 > > were style (80 character limit, bad indentations of 4 instead of
 > 2,
 > 
 > Most of android is 4 spaces for indents and 100 char limits.
 > 
 > > missing docstring, implicit True/False condition (e.g. not len),
 > > space after [, space before ] and " delimiter inconsistencies).
 > >
 > > Having troubles finding a free reviewer that wants to claim
 > google
 > > py readability.
 > 
 > Ill take a look through the linter, but its likely reporting a
 > style that
 > it wants, not a style that I want.
 > 
 > IIRC, I loaded this into PyDev (Eclipse) and formatted it.

I used normal pylint, and yapf (https://github.com/google/yapf) with the format style set to Google. I also found this style guide here:

https://google.github.io/styleguide/pyguide.html

and

http://sphinxcontrib-napoleon.readthedocs.org/en/latest/example_google.html


The only thing I may be deviating from is the simple accessor methods on the classes are simple and should likely just be a variable.


I got the errors down, however it still reports some, but the tool is wrong. The overall code score from pylint is 9.4ish (depending on which patch).

I think this is now the most documented tool in all of build :-P (likely very wrong with that statement).

Patch-set: 4
Label: Verified=0