fs_config: limit characters for AID_<name> sections
When generating files restrict the characters allowed in
AID_<name> to upercase, numbers, and underscores.

This detects errors ahead of time for generated C files as
well as handles seperation characters for passwd/group files.

This also lends itself to automatic detection of collisions
on friendly names, since freindly names are the lowercase
version of <name>.

Change-Id: Icc6c43f79462682d1ba6c2b824c092de11bbf9ab
Signed-off-by: William Roberts <william.c.roberts@intel.com>