Update patch set 1
Patch Set 1:

> Ying: Why shouldn't we just enable checkbuild for the second arch?
 > It'll take longer, but that's what people are expecting checkbuild
 > to do I think (I didn't realize this exception, and Soong's
 > checkbuild includes everything)

Yes, the checkbuild time was the main concern. The idea was that the 32-bit should already be checked in 32-bit-only build. 
In my tests, it took about 15-20% longer to checkbuild everything.
I don't have strong opinion.
Remove the exception then we don't need this CL. Looks fine to me.

Patch-set: 1