Update patch set 1
Patch Set 1:

> Ian,
 > 
 > I had this problem too with warnings from clang-tidy and static
 > analyzer.  Most of the warnings were from external or vendor
 > projects that we might not fix all of them in a short time. But
 > some warnings are easy to fix and we want
 > to fix one warning kind at a time.
 
I'd say this is probably less the case with Java code, compared to C++ code, but the point is taken.
 
 > Limiting dump of warnings to 250 in dumpcategory still has a few
 > problems.
 > (0) We should put this under some flag, so we can see all warnings
 > in a nightly or special build report.

With this change, the number of warnings is still displayed. The list of all instances of the warnings is cut at 250.

 > (1) It is too late after hundreds of thousands of warnings were
 > dumped into the build.log file, and it took a long time to sort
 > those warnings.

I'm not too fussed at how long warn.py takes to run.

 > (2) It cuts off the warnings we want to see.  Warnings for files
 > under "frameworks" and "system" projects are sorted after
 > "external" projects. We will see 250 warnings up to some "external"
 > project and not those from "frameworks" and "system".
 
250 could equally be 500. I agree that having some kind of sorting/prioritization would be good. I'd add that lowering the priority of files with test and cts in the name would make sense.
 
 > Maybe warn.py needs more flags to choose which projects and which
 > warning kinds to show, and/or better structured output in html to
 > select/hide different warning kinds.
 
I think keeping warn.py simple also has virtues.

 > For an easy fix to avoid bloating up build.log, I have enabled only
 > some clang-tidy warnings as default and disabled some for external
 > projects, see build/core/clang/tidy.mk.  I can enable more warnings
 > through environment variables for my own dev/test builds. Could
 > this work for the "Error Prone" warnings?

The only really bad one is MissingOverrides, we did a round of removing noisy warnings, but MissingOverrides is useful in catching real errors, so we held onto it.

All of this said, I think it remains a bug to generate a HTML file that a browser struggles to load. I think cutting the warnings at some large number makes sense (500?). Longer term a better fix would be to output the warnings in a more condensed form after some number - say just the first instance of a warning in a file.

Patch-set: 1