Update patch set 1
Patch Set 1:

> Maybe the best is to limit the number of warnings from the original
 > source. Could "Error Prone" limit warnings from selected files?
 > How else can people work on fixing one kind of warnings from one
 > source file at a time?
 > 
 > If that does not work and this case is only for non-default build
 > configuration, maybe a non-default flag is enough for those runs?
 > 
 > Some clang-tidy warnings have the same problem and I don't see any
 > solution except disabling those warnings in default configuration.
 > 
 > Current enabled warnings from clang-tidy and static analyzer are
 > not many in one file, but there are too many files with the same
 > kind of warning. There could be up to tens of thousands in one
 > kind. That seems to work fine for build.log and browser. So a
 > default limit of 1000 total for one kind or just a few per file
 > seems too early for C/C++ nightly build warnings.
 > 
 > Another simple solution for human readers is to add in the output
 > html file some buttons to hide/expand each warning list. Warnings
 > with hundreds of cases could be in a hidden state by default.

So let's change the number to 1000. With this change (and the number being 1000) you will still see that there were >1000 failures of a particular kind. Within that failure, after 1000 of those failures were added to the HTML table, the table will stop. I don't think people will scroll through >1000 cases of a particular error, but having the HTML for all of those errors makes the warnings.html large and unmanageable. It seems the downside of this approach is that when you try to find a particular file name, if it falls into the >1000 camp, you won't see it. I think I care less about that case (we already have 1000 instances of the warning) than the unmanageably large HTML file.

Patch-set: 1