Clean installed files that would no longer be installed by default
This is an 80% solution that covers the common cases of modules being
removed from the default install set of PRODUCT_PACKAGES /
PRODUCT_COPY_FILES. It's possible that it misses some dependencies, and
extra installed files (like symlinks), but dealing with those would
significantly complicate this.

Ideally we'd be keeping track of what the user explicitly asked to be
installed more than the default, and clean those as necessary as well,
but that's even more involved, since that's handled in ninja.

Bug: 29160799
Change-Id: I0b2c574865f93e072fb5fc50cbb8848ba8d12b34