Fix root_filesystem_config.txt generation
The first line returned by awk is empty. In consequence, when
fs_config computes its mode it will consider this line as a file
because no trailing slash) and affect the default android_files mode
which is 0644.
The mode for the root directory should be the default android_dirs
mode 0755. Add a sed command to replace the empty line by a slash
before exectuing fs_config.

Change-Id: I9f33f6fcf4be05c31914db898e65c92b0a611518
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>