Support creation of EFI partition image using make_efi_image.
make_efi_image, as discussed in
https://android-review.googlesource.com/#/c/243590/, allows for the
creation of EFI images to be flashed on hard disk images.  Integration
of this script is done here so that if EFI_INPUT_FILE_PATHS is non-empty,
the build system system will call the script, passing EFI_IMAGE_SIZE,
$(ANDROID_OUT)/EFI.img, and EFI_INPUT_FILE_PATHS as arguments.

BUG=29123391
TEST=Manually observed contents of EFI.img. Checked contents of zipfile
in out/ to verify EFI.img was contained.

Change-Id: I4e3c641cc286f46f6c578413bdfe2794267fc84a