Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

I think the reason I wrote it this way was that I was hoping we'd have things like msan and tsan much sooner. We have some unit tests that currently have `LOCAL_SANITIZE := address` always set, and that will break if we have `SANITIZE_TARGET := thread`.

I don't think that's a problem that we can encounter yet, so I'm okay with this going in and us dealing with that in soong where we can perform those checks a bit easier.

Patch-set: 2
Label: Code-Review=+2