Verify prebuilt libraries have SONAME, correct arch
We can't link against prebuilt libraries that don't have a correct
SONAME anymore, since we're no longer linking with -L out of the temporary
obj/lib path, but using the full path. If the SONAME isn't set, the full
path on the build server would be used in DT_NEEDED.

This will also warn on extra/missing DT_NEEDED entries compared to the
contents of LOCAL_SHARED_LIBRARIES to ensure that the dependencies are
correct.

Test: m verify_all_prebuilt_libraries (aosp and internal)
Change-Id: Id8990b674a2a7343976656c0380d3900cd387ccf