Update patch set 1
Patch Set 1:

> Do we know what broke?

Kind of. PIC has never worked for MIPS, possibly because of what we just found. We found that zygote hangs in an infinite loop in here:

15: short libcore.io.Memory.peekShort(long, boolean) (dex_method_idx=18363)
...
core_spill_mask: 0x80000001 (r0, r31)
...
      0x008ba6b0: 8fa2000c      lw      r2, +12(r29)
      0x008ba6b4: 04100000      bltzal  r0, 0x008ba6b8  ; +4
      0x008ba6b8: 3c03fe6d      lui     r3, 0xfe6d0000
      0x008ba6bc: 34631784      ori     r3, r3, 6020
      0x008ba6c0: 007f1821      addu    r3, r3, r31
      0x008ba6c4: 80c10001      lb      r1, +1(r6)
      0x008ba6c8: 80c30000      lb      r3, +0(r6)
      0x008ba6cc: 7c23fa04      ins     r3, r1, 8, 24
      0x008ba6d0: 10020003      beq     r0, r2, 0x008ba6e0  ; +16
      0x008ba6d4: 00000000      nop     
      0x008ba6d8: 7c0318a0      wsbh    r3, r3
      0x008ba6dc: 7c031e20      seh     r3, r3
      0x008ba6e0: 00601025      or      r2, r3, r0
      0x008ba6e4: 03e00009      jr      r31
      0x008ba6e8: 00000000      nop     

The NAL(BLTZAL) instruction changes RA to point to the ORI instruction, which is how we get into an infinite loop.

This sequence of four instructions (NAL, LUI, ORI, ADDU) is there to implement PC-relative addressing.

However, it's odd that the resultant value from these instructions (V1(R3)=PC+constant) is not used. And it's odd that core_spill_mask has more than just RA in it and yet there's no frame where RA would be saved.

IOW, there's a dex2oat bug either in generating this unused sequence or in not creating a stack frame with RA preserved in it. Or both.

When PIC is off, the four instructions aren't there and core_spill_mask contains only RA and all is well.

Patch-set: 1