Remove --ignore_dirty=$(OUT_DIR)/% from Kati flags
We really should be re-running Kati when a file in here changes, since
it may actually affect the result. The general problem we've run into is
when there are two kati output files that are clean according to the
current rules, but if they actually ran we'd force an installclean or
generated java file cleanup.

This also had the side-effect of not running shell commands that touched
$OUT_DIR, like writing files using $(shell echo ... >out/...), now those
may be run.

This will cause Kati to re-read the makefiles more often -- twice when a
CleanSpec.mk changes, twice when switching products on the same device,
twice when the generated java files changes. Basically every time we
read the old state, do something about it, and write out the new state,
we need to rerun the next time so we aren't continuously cleaning up
what changed. I'm fixing the generated java cleanups in the next commit
by moving the reading of the old state out into the wrapper so that it
runs every time, but that's more difficult with the CleanSpec and build
config changes.

Bug: 30947985
Test: Manual, switching between configurations; enable --regen_debug
Change-Id: I888f9f8295f318f8214b2c79525cf94f457a5a8b