Fix link_type checking
This was printing "error:", but not actually triggering an error.
Instead of trying to write a single line bash script to handle this,
move the actual check into python. This allows us to print all of the
errors for a single module before triggering the failure.

Also updates the warning format and the warn.py script to properly parse
this warning. Assigning a LOW priority, since many of the failures are
false positives due to the lack of LOCAL_SDK_VERSION marking on
prebuilts.

Test: Check warnings being printed, add error and make sure it errors.
Change-Id: I333e418c9a4511b7c7e826891ae481da08fbf6f9