Update patch set 3
Patch Set 3:

Let me confirm my understanding is correct or not.

If I understand correctly, making GOMACC_PATH global or not might not be the point.  The reason why you think this change is bad is that enabling / disabling USE_GOMA won't cause full rebuild if this change is submitted as-is.

Then, making ninja or something else rebuild all on certain specified environment change could be yet another approach to meet your requirement.

Is my understanding correct?

Patch-set: 3