Update patch set 3
Patch Set 3:

> The right way to do this would be to set GOMACC_PATH individually for each command that uses it.

If possible, I would like to go with that way.  However, there is some unclear part for me to do that.

I assume you might be one of the soong developers, and soong might be a program that generate build.ninja, and it might be possible to set GOMACC_PATH individually for each command that uses it.

If I pass a gomacc path as a command line argument for a wrapper, what is a preferred way to make soong know the gomacc path?

Also, it it possible for soong to detect a compiler is a wrapper or an ELF/Mach-O binary?  If yes, is it possible to change the command line generated?
e.g. if a compiler is a wrapper,
clang --gomacc /some/where/gomacc <original args>
otherwise,
/some/where/gomacc clang <original args>

Patch-set: 3