Generate SHA-1 build ID for host-generated *.oat files (2/2).
For host-generated *.oat files, generate a SHA-1 build ID based on the
file content and write it to .note.gnu.build-id ELF section.  This
should allow various developer tools like profilers correlate the data
captured for files like boot.oat on the device with the corresponding
known version of the file during an offline analysis.

Bug: 31292208
Test: Verified that boot.oat contains the build ID section now.
Test: make test-art-host
Change-Id: Iee48e437009488f20995e73868d5059d349701c0