Update patch set 3
Patch Set 3:

> There's some warnings internally that shouldn't be there, and I was
 > going to investigate. But it's probably better to have more
 > warnings and some errors rather than errors that don't do anything,
 > so I'll investigate and do a follow-on change. (It looks like tests
 > installed in /data linking to /vendor code aren't well handled)

I really think this is not a sustainable strategy in enabling new warnings like this. In internal master this change added 126 new warnings. We should really be fixing these issues before enabling the warning otherwise we get back to the situation where we have thousands of warnings and nobody is looking at them anymore due to warnings being lost in noise.

Patch-set: 3