Update patch set 1
Patch Set 1:

Shouldn't we just be putting the other warnings into an "unknown" category?

Ah, I see from the source in the bug that we're picking "warning:" inside a bash command that ninja is running, not from the output itself. Avoiding "[" at the beginning of the line may be a better approach.

The SyntaxError problem means that we should be escaping these strings...

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0