Update patch set 2
Patch Set 2:

> What configurations are warning here? I don't see any warnings on my local builds.

These are static analyzer warnings. You'll need to compile with WITH_TIDY=1 and WITH_STATIC_ANALYZER=1 in order to see them.

> so I'm worried that we're not really initializing these elsewhere

Your hunch was right. :) we had a few paths where compressFpToFp would fail and return NO_ERROR (...and never assign to `crc`), so the static analyzer was partially complaining about that.

The other part of the complaint was that -DNDEBUG is given, so the asserts here get compiled into nothing, so the analyzer thinks compressionMethod can be something other than kCompressDeflated or kCompressStored. We could apply a local fix for this with some ifdef magic, but I think we might be able to fix it more cleanly in Bionic.

Patch-set: 2