Make DROIDDOC rules optionally included in 'docs' target
The default is still to be included but targets can opt out
via the LOCAL_EXCLUDE_FROM_ALL_DOCS variable. This is being
done to allow libcore-docs to be excluded: it is only supposed
to be a quick target for libcore developers to check docs
without doing the full docs build, but it looks like it was
slowing down the full docs build as a side effect.

Bug: 32128132
Test: Dirtied something in libcore, ran 'make docs', confirmed
that 'make libcore-docs' did work, but then running
'make libcore-docs' a second time had nothing to do.

Change-Id: I095c180358f92744f1c875dd4fb285ad9867ab84