Update patch set 2
Patch Set 2:

If you look at the companion changes, the rules are in userdebug_or_eng, but moved out the type line.

Because the type line is moved out and can not be used to give us a signal of errant rules, we decided to move out the neverallow rules too.

Patch-set: 2