Check includes for modules with LOCAL_USE_VNDK and generate warnings
Warnings can be fixed by removing specific includes and adding libs
to LOCAL_SHARED_LIBRARIES.

Even after the fix, a circular dependency might be reported. That means
the framework module depends on a vendor module. So the interface will
need to be HIDL-ized.

Test: Enable BOARD_VNDK_VERSION in BoardConfig.mk to turn on
LOCAL_USE_VNDK and see the warnings
Change-Id: Ie0de668f175ccb551ca1ab733989d200031e6a1c