Update patch set 1
Patch Set 1:

Generally like the idea of speeding up end-user builds with this, but: Today we often get complaints that TreeHugger is slow and flaky because (1) they built the target locally and it only took 5 minutes; and (2) their change worked locally but failed on TreeHugger. How do we prevent that mindset from getting worse with this change? Today, at least, users don't really know to run checkbuild or build with tests.

Patch-set: 1