Don't run check commands unless we're in checkbuild mode.
checkbuild commands should only be built in checkbuild mode -- not all
the time. Fix this.

This fixes a bug where jack check commands are being run in every build
(and slowing down the build noticeably). Otherwise, it has little
impact, because for most modules, the checkbuild command is identical to
the regular build command.

Bug: 30439935
Test: Verify faster builds in presubmit, examine ninja logs.
Change-Id: I166b0e0913721dafebe659ae1da846529ae6051d