Update patch set 1
Patch Set 1:

Correction 6 tests failed.

I'm trying to understand if I missed something in fsconfig2-fixups or how this broke. The closest I can find in the series that was squashed into this is: https://android-review.googlesource.com/#/c/313670/10/tools/fs_config/fs_config_generator.py

Which that change seems to correctly have the _ removed. Was this accidentally introduced in the squash? From what I can tell diffing the head of fsconfig2-fixups with the revert revert squash commit is that not everything made it into the squash commit.

Notably some yapf formatting fixes and pylint changes were never in there:

diff:
-    _OEM_RANGE = re.compile('%sOEM_RESERVED_[0-9]*_{0,1}(%s|%s)' %
+    _OEM_RANGE = re.compile('%s_OEM_RESERVED_[0-9]*_{0,1}(%s|%s)' %

                 """Creates an error message with the current parsing state."""
-                # pylint: disable=cell-var-from-loop
                 return 'Error "{}" in file: "{}" on line: {}'.format(

-                    sys.exit(
-                        error_message('{} for "{}"'.format(exception,
-                                                           identifier)))
+                    sys.exit(error_message(
+                        '{} for "{}"'.format(exception, identifier)))

Patch-set: 1