Update patch set 1
Patch Set 1:

The LOCAL_REQUIRED_MODULES in test modules is currently unused (or if used it is useless). This is because a clean "make dist tests" will not include those modules in neither the system image nor the test .zip... that's why I'm re-using LOCAL_REQUIRED_MODULES which already does what you'd expect when building.

The test systems will do the same that they do with the LOCAL_PICKUP_FILES for example, except that in this case you specify a module name instead of a source filename. The test systems will unpack that directory and run only the test executable; but the test executable could depend on other files in the same directory. 

Examples:
1. LOCAL_REQUIRED_MODULES includes another executable binary module (other_binary) whose destination directory is the same as your test module (/data/nativetest/mytest/other_binary, right next to /data/nativetest/mytest/mytest). At runtime, mytest can call "other_binary" from the same directory.

2. Include a .so library module. The current logic will include them in the /lib directory inside the .zip file (DATA/lib/foo.so) . You see this already happening with some "art" libraries. When running the test, you can add an extra LD_LIBRARY_PATH to make it find the /data/lib/foo.so. I would expect this to be the default, but that's a separated discussion.

I added $(warning) lines to the .mk to manually verify the list of modules before and after the expansion and I ran "make dist tests" and verified that the continuous_native_test.zip file contains the dependencies for a target that includes a test module with LOCAL_REQUIRE_MODULES (update_engine_unittests)

Patch-set: 1