install *.so in different paths for their types
Shared libraries are now installed to different directories depending on
their types.

* VNDK libraries: /system/lib/vndk
* VNDK-ext libraries: /system/lib/vndk-ext
* Framework-only libraries: /system/lib/framework
* NDK libraries: /system/lib
* Vendor-only libraries: /vendor/lib
* Same-process HALs: /vendor/lib/sameprocess

However, if LOCAL_MODULE_PATH is explicitly set, then it is respected,
with a warning message. Module owners are highly encouraged to
investigate the warnings and use alternatives to LOCAL_MODULE_PATH;
combination of LOCAL_[PROPRIETARY|OEM|ODM]_MODULE, LOCAL_MODULE_CLASS
and LOCAL_RELATIVE_PATH will cover most of the cases.

Furthermore, for each shared libraries whose path is changed, a symolic
link from the original path to the new path is *temporarily* generated.
e.g. /system/lib/libgui.so -> framework/libgui.so. This is
to prevent sudden breakage of the code expecting the lib from the old
path. This symbolic links will eventually be removed before O launch
(b/34917183).

In addition, a new flag LOCAL_EXTENDS_MODULE is added. This is used to
specify the module that this module extends. Currently, this is only
used to extend a vndk lib. For example, in order to extend a vndk
library libFoo:

LOCAL_MODULE := libFoo
LOCAL_SRC_FILES := path/to/src
...
include $(BUILD_SHARED_LIBRARY)
----------------------
LOCAL_MODULE := libFoo-extended
LOCAL_SRC_FILES := path/to/extended-src
...
LOCAL_EXTENDS_MODULE := libFoo
include $(BUILD_SHARED_LIBRARY)

libFoo will be installed as /system/lib/vndk/libFoo.so and
libFoo-extended will be installed as /system/lib/vndk-ext/libFoo.so.
Note that file name of the latter is libFoo.so, not libFoo-extended.so:
file name of an extending module is automatically set to that of the
extended module.

Finally, BOARD_SAME_PROCESS_HAL_DEPS is added. It contains the list of
shared libraries implementing the same-process HALs and its internal sub
libraries. This is designed to be defined in BoardConfig.mk

Bug: 33681361
Test: build & run. Libraries must be in the correct directories.
Symlinks from the old path to the new path must exist.

Change-Id: I46130aac09ae65400fd4d0abbc2e12dddd154fb1