Increase USER length limit from 6 to 8 in BUILD_NUMBER.
Long user names are truncated to make it more likely that the entire
BUILD_NUMBER string will fit. Many legacy systems (eg. NIS, to name one
well-known) limit user names to 8 characters so if the Android build
system only truncates user names longer than that then the number of
user names that will need to be truncated can be minimized. The user
name is only a small part of this string so this possible addition of
2 additional characters won't significantly increase the risk that the
string becomes too long.

Test: make
Change-Id: I76ee1be990ed6497ec5d52c094804a0a61a4a67d