Generate SHA-1 for each block and add to target file
Generate SHA-1 for the blocks on the care map and save the data in
META/block_sha1.txt. The file will be helpful for debugging as it has
all the correct hashes.

Also this file is ~11M for bullhead, too large for an OTA package.
The option to reduce the size includes using CRC instead of SHA-1, and
calculating hashes for a larger cluster of blocks (for example, one hash
every 16 blocks).

Bug: 21124445
Test: block_sha1.txt generates successfully for bullhead
Change-Id: I2ef40b591b6e99b63fd95b4b815736e47492a520