Generate SHA-1 for each block and add to target file
Generate SHA-1 for the blocks on the care map and save the data in
META/block_sha1.txt. The file will be helpful for debug as it has all
the correct hashes.

Also this file is ~11M for bullhead, too large for an OTA package.
Some thoughts to reduce the size includes using CRC32 instead of SHA-1,
and to calculate hashes for a larger cluster of blocks (for example,
one hash every 16 blocks).

Bug: 21124445
Test: block_sha1.txt generates successfully for bullhead
Change-Id: I2ef40b591b6e99b63fd95b4b815736e47492a520