Update patch set 2
Patch Set 2: -Code-Review

Actually this is not intended for system apps...for these there is no need. But for apps that are not system. OEMs sometimes add apps to /system or /vendor images that will not be installed by Package Manager during boot scan, but it will be installed by some installer service as user app.

Patch-set: 2
Reviewer: Gerrit User 1070179 <1070179@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0