Update patch set 1
Patch Set 1: Presubmit-Ready+1

No, old files left behind during incremental builds are not a problem. Before this commit, there is only /sepolicy and /system/bin/secilc. After this commit, /system/bin/secilc disappears in clean builds (but may linger in incremental ones) on all devices where PRODUCT_FULL_TREBLE is not set to "true". secilc is not used by init on those devices, so, it's not an issue. On PRODUCT_FULL_TREBLE devices, after this commit, /sepolicy  disappears in clenan builds but might linger in incremental builds. This is not an issue because after this commit /sepolicy is not used on these devices.

Patch-set: 1
Reviewer: Gerrit User 1018773 <1018773@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1